backend/requirements.txt
fastapi==0.115.0
httpx==0.27.2
paho-mqtt==2.1.0
uvicorn==0.30.6
